requirements.txt
Django==1.8.2
psycopg2==2.6
django-grappelli==2.6.4
djangorestframework==3.3.3
Pillow==2.6.1

